Contract reader/writer refactoring
