+++ apps/block_scout_web/assets/js/lib/smart_contract/functions.js
@@ -1,7 +1,6 @@
 import $ from 'jquery'
-import { props } from 'eth-net-props'
-import { walletEnabled, connectToWallet, getCurrentAccount, shouldHideConnectButton } from './write.js'
-import { openErrorModal, openWarningModal, openSuccessModal, openModalWithMessage } from '../modals.js'
+import { getContractABI, getMethodInputs, prepareMethodArgs } from './common_helpers'
+import { walletEnabled, connectToWallet, shouldHideConnectButton, callMethod } from './write'
 import '../../pages/address'
 
 const loadFunctions = (element) => {
@@ -135,181 +134,6 @@ const readWriteFunction = (element) => {
         .then((isWalletEnabled) => callMethod(isWalletEnabled, $functionInputs, explorerChainId, $form, functionName, $element))
     }
   })
-}
-
-function callMethod (isWalletEnabled, $functionInputs, explorerChainId, $form, functionName, $element) {
-  if (!isWalletEnabled) {
-    const warningMsg = 'You haven\'t approved the reading of account list from your MetaMask or MetaMask/Nifty wallet is locked or is not installed.'
-    return openWarningModal('Unauthorized', warningMsg)
-  }
-  const contractAbi = getContractABI($form)
-  const inputs = getMethodInputs(contractAbi, functionName)
-
-  const $functionInputsExceptTxValue = $functionInputs.filter(':not([tx-value])')
-  const args = prepareMethodArgs($functionInputsExceptTxValue, inputs)
-
-  const txValue = getTxValue($functionInputs)
-  const contractAddress = $form.data('contract-address')
-
-  const { chainId: walletChainIdHex } = window.ethereum
-  compareChainIDs(explorerChainId, walletChainIdHex)
-    .then(currentAccount => {
-      if (functionName) {
-        const TargetContract = new window.web3.eth.Contract(contractAbi, contractAddress)
-        const sendParams = { from: currentAccount, value: txValue || 0 }
-        const methodToCall = TargetContract.methods[functionName](...args).send(sendParams)
-        methodToCall
-          .on('error', function (error) {
-            openErrorModal(`Error in sending transaction for method "${functionName}"`, formatError(error), false)
-          })
-          .on('transactionHash', function (txHash) {
-            onTransactionHash(txHash, $element, functionName)
-          })
-      } else {
-        const txParams = {
-          from: currentAccount,
-          to: contractAddress,
-          value: txValue || 0
-        }
-        window.ethereum.request({
-          method: 'eth_sendTransaction',
-          params: [txParams]
-        })
-          .then(function (txHash) {
-            onTransactionHash(txHash, $element, functionName)
-          })
-          .catch(function (error) {
-            openErrorModal('Error in sending transaction for fallback method', formatError(error), false)
-          })
-      }
-    })
-    .catch(error => {
-      openWarningModal('Unauthorized', formatError(error))
-    })
-}
-
-function getMethodInputs (contractAbi, functionName) {
-  const functionAbi = contractAbi.find(abi =>
-    abi.name === functionName
-  )
-  return functionAbi && functionAbi.inputs
-}
-
-function prepareMethodArgs ($functionInputs, inputs) {
-  return $.map($functionInputs, (element, ind) => {
-    const inputValue = $(element).val()
-    const inputType = inputs[ind] && inputs[ind].type
-    let sanitizedInputValue
-    sanitizedInputValue = replaceSpaces(inputValue, inputType)
-    sanitizedInputValue = replaceDoubleQuotes(sanitizedInputValue, inputType)
-    if (isArrayInputType(inputType)) {
-      if (sanitizedInputValue === '') {
-        return [[]]
-      } else {
-        if (sanitizedInputValue.startsWith('[') && sanitizedInputValue.endsWith(']')) {
-          sanitizedInputValue = sanitizedInputValue.substring(1, sanitizedInputValue.length - 1)
-        }
-        const inputValueElements = sanitizedInputValue.split(',')
-        const sanitizedInputValueElements = inputValueElements.map(elementValue => {
-          const elementInputType = inputType.split('[')[0]
-          return replaceDoubleQuotes(elementValue, elementInputType)
-        })
-        return [sanitizedInputValueElements]
-      }
-    } else { return sanitizedInputValue }
-  })
-}
-
-function isArrayInputType (inputType) {
-  return inputType && inputType.includes('[') && inputType.includes(']')
-}
-
-function isAddressInputType (inputType) {
-  return inputType.includes('address')
-}
-
-function isUintInputType (inputType) {
-  return inputType.includes('int') && !inputType.includes('[')
-}
-
-function isStringInputType (inputType) {
-  return inputType.includes('string') && !inputType.includes('[')
-}
-
-function isNonSpaceInputType (inputType) {
-  return isAddressInputType(inputType) || inputType.includes('int') || inputType.includes('bool')
-}
-
-function getTxValue ($functionInputs) {
-  const WEI_MULTIPLIER = 10 ** 18
-  const $txValue = $functionInputs.filter('[tx-value]:first')
-  const txValue = $txValue && $txValue.val() && parseFloat($txValue.val()) * WEI_MULTIPLIER
-  const txValueStr = txValue && txValue.toString(16)
-  return txValueStr
-}
-
-function getContractABI ($form) {
-  const implementationAbi = $form.data('implementation-abi')
-  const parentAbi = $form.data('contract-abi')
-  const $parent = $('[data-smart-contract-functions]')
-  const contractType = $parent.data('type')
-  const contractAbi = contractType === 'proxy' ? implementationAbi : parentAbi
-  return contractAbi
-}
-
-function compareChainIDs (explorerChainId, walletChainIdHex) {
-  if (explorerChainId !== parseInt(walletChainIdHex)) {
-    const networkDisplayNameFromWallet = props.getNetworkDisplayName(walletChainIdHex)
-    const networkDisplayName = props.getNetworkDisplayName(explorerChainId)
-    const errorMsg = `You connected to ${networkDisplayNameFromWallet} chain in the wallet, but the current instance of Blockscout is for ${networkDisplayName} chain`
-    return Promise.reject(new Error(errorMsg))
-  } else {
-    return getCurrentAccount()
-  }
-}
-
-function onTransactionHash (txHash, $element, functionName) {
-  openModalWithMessage($element.find('#pending-contract-write'), true, txHash)
-  const getTxReceipt = (txHash) => {
-    window.ethereum.request({
-      method: 'eth_getTransactionReceipt',
-      params: [txHash]
-    })
-      .then(txReceipt => {
-        if (txReceipt) {
-          const successMsg = `Successfully sent <a href="/tx/${txHash}">transaction</a> for method "${functionName}"`
-          openSuccessModal('Success', successMsg)
-          clearInterval(txReceiptPollingIntervalId)
-        }
-      })
-  }
-  const txReceiptPollingIntervalId = setInterval(() => { getTxReceipt(txHash) }, 5 * 1000)
-}
-
-function replaceSpaces (value, type) {
-  if (isNonSpaceInputType(type)) {
-    return value.replace(/\s/g, '')
-  } else {
-    return value
-  }
-}
-
-function replaceDoubleQuotes (value, type) {
-  if (isAddressInputType(type) || isUintInputType(type) || isStringInputType(type)) {
-    if (typeof value.replaceAll === 'function') {
-      return value.replaceAll('"', '')
-    } else {
-      return value.replace(/"/g, '')
-    }
-  } else {
-    return value
-  }
-}
-
-const formatError = (error) => {
-  let { message } = error
-  message = message && message.split('Error: ').length > 1 ? message.split('Error: ')[1] : message
-  return message
 }
 
 const container = $('[data-smart-contract-functions]')
