+++ apps/block_scout_web/assets/js/lib/smart_contract/write.js
@@ -1,4 +1,7 @@
 import Web3 from 'web3'
+import { props } from 'eth-net-props'
+import { openErrorModal, openWarningModal, openSuccessModal, openModalWithMessage } from '../modals'
+import { getContractABI, getMethodInputs, prepareMethodArgs } from './common_helpers'
 
 export const walletEnabled = () => {
   return new Promise((resolve) => {
@@ -79,4 +82,98 @@ export const shouldHideConnectButton = () => {
       resolve({ shouldHide: false })
     }
   })
+}
+
+export function callMethod (isWalletEnabled, $functionInputs, explorerChainId, $form, functionName, $element) {
+  if (!isWalletEnabled) {
+    const warningMsg = 'You haven\'t approved the reading of account list from your MetaMask or MetaMask/Nifty wallet is locked or is not installed.'
+    return openWarningModal('Unauthorized', warningMsg)
+  }
+  const contractAbi = getContractABI($form)
+  const inputs = getMethodInputs(contractAbi, functionName)
+
+  const $functionInputsExceptTxValue = $functionInputs.filter(':not([tx-value])')
+  const args = prepareMethodArgs($functionInputsExceptTxValue, inputs)
+
+  const txValue = getTxValue($functionInputs)
+  const contractAddress = $form.data('contract-address')
+
+  const { chainId: walletChainIdHex } = window.ethereum
+  compareChainIDs(explorerChainId, walletChainIdHex)
+    .then(currentAccount => {
+      if (functionName) {
+        const TargetContract = new window.web3.eth.Contract(contractAbi, contractAddress)
+        const sendParams = { from: currentAccount, value: txValue || 0 }
+        const methodToCall = TargetContract.methods[functionName](...args).send(sendParams)
+        methodToCall
+          .on('error', function (error) {
+            openErrorModal(`Error in sending transaction for method "${functionName}"`, formatError(error), false)
+          })
+          .on('transactionHash', function (txHash) {
+            onTransactionHash(txHash, $element, functionName)
+          })
+      } else {
+        const txParams = {
+          from: currentAccount,
+          to: contractAddress,
+          value: txValue || 0
+        }
+        window.ethereum.request({
+          method: 'eth_sendTransaction',
+          params: [txParams]
+        })
+          .then(function (txHash) {
+            onTransactionHash(txHash, $element, functionName)
+          })
+          .catch(function (error) {
+            openErrorModal('Error in sending transaction for fallback method', formatError(error), false)
+          })
+      }
+    })
+    .catch(error => {
+      openWarningModal('Unauthorized', formatError(error))
+    })
+}
+
+function onTransactionHash (txHash, $element, functionName) {
+  openModalWithMessage($element.find('#pending-contract-write'), true, txHash)
+  const getTxReceipt = (txHash) => {
+    window.ethereum.request({
+      method: 'eth_getTransactionReceipt',
+      params: [txHash]
+    })
+      .then(txReceipt => {
+        if (txReceipt) {
+          const successMsg = `Successfully sent <a href="/tx/${txHash}">transaction</a> for method "${functionName}"`
+          openSuccessModal('Success', successMsg)
+          clearInterval(txReceiptPollingIntervalId)
+        }
+      })
+  }
+  const txReceiptPollingIntervalId = setInterval(() => { getTxReceipt(txHash) }, 5 * 1000)
+}
+
+const formatError = (error) => {
+  let { message } = error
+  message = message && message.split('Error: ').length > 1 ? message.split('Error: ')[1] : message
+  return message
+}
+
+function getTxValue ($functionInputs) {
+  const WEI_MULTIPLIER = 10 ** 18
+  const $txValue = $functionInputs.filter('[tx-value]:first')
+  const txValue = $txValue && $txValue.val() && parseFloat($txValue.val()) * WEI_MULTIPLIER
+  const txValueStr = txValue && txValue.toString(16)
+  return txValueStr
+}
+
+function compareChainIDs (explorerChainId, walletChainIdHex) {
+  if (explorerChainId !== parseInt(walletChainIdHex)) {
+    const networkDisplayNameFromWallet = props.getNetworkDisplayName(walletChainIdHex)
+    const networkDisplayName = props.getNetworkDisplayName(explorerChainId)
+    const errorMsg = `You connected to ${networkDisplayNameFromWallet} chain in the wallet, but the current instance of Blockscout is for ${networkDisplayName} chain`
+    return Promise.reject(new Error(errorMsg))
+  } else {
+    return getCurrentAccount()
+  }
 }
