+++ apps/block_scout_web/lib/block_scout_web/templates/address_write_contract/index.html.eex
@@ -16,5 +16,5 @@
     </div>
   </div>
   <script defer data-cfasync="false" src="<%= static_path(@conn, "/js/smart-contract-helpers.js") %>"></script>
-  <script defer data-cfasync="false" src="<%= static_path(@conn, "/js/write_contract.js") %>"></script>
+  <script defer data-cfasync="false" src="<%= static_path(@conn, "/js/write-contract.js") %>"></script>
 </section>
