+++ apps/block_scout_web/lib/block_scout_web/templates/smart_contract/_connect_container.html.eex
@@ -1,0 +1,13 @@
+<div connect-to class="connect-container">
+    <span style="margin-top: -2px;" class="mr-2">
+      <%= render BlockScoutWeb.IconsView, "_inactive_icon.html" %>
+    </span>
+     <h2 style="margin-top: -2px;">Disconnected</h2>
+    <button connect-metamask class="button btn-line ml-4">Connect to Metamask</button>
+</div>
+<div connected-to class="connect-container hidden">
+    <span style="margin-top: -2px;" class="mr-2">
+      <%= render BlockScoutWeb.IconsView, "_active_icon.html" %>
+    </span>
+    <h2 style="margin-top: -2px;">Connected to</h2><h3 connected-to-address class="ml-2"></h3>
+</div>
