+++ apps/block_scout_web/lib/block_scout_web/templates/smart_contract/_functions.html.eex
@@ -12,19 +12,7 @@ to: address_contract_path(@conn, :index, metadata_for_verification.address_hash)
   <% end %>
 <% end %>
 <%= if @action== "write" do %>
-  <div connect-to class="connect-container">
-    <span style="margin-top: -2px;" class="mr-2">
-      <%= render BlockScoutWeb.IconsView, "_inactive_icon.html" %>
-    </span>
-     <h2 style="margin-top: -2px;">Disconnected</h2>
-    <button connect-metamask class="button btn-line ml-4">Connect to Metamask</button>
-  </div>
-  <div connected-to class="connect-container hidden">
-    <span style="margin-top: -2px;" class="mr-2">
-      <%= render BlockScoutWeb.IconsView, "_active_icon.html" %>
-    </span>
-    <h2 style="margin-top: -2px;">Connected to</h2><h3 connected-to-address class="ml-2"></h3>
-  </div>
+  <%= render BlockScoutWeb.SmartContractView, "_connect_container.html" %>
 <% end %>
 <%= if @contract_type == "proxy" do %>
 <div class="implementation-container">
@@ -52,7 +40,7 @@ to: address_contract_path(@conn, :index, metadata_for_verification.address_hash)
           end
         %>
         <%= render BlockScoutWeb.SmartContractView, "_pending_contract_write.html" %>
-        <form class="form-inline" data-function-form data-action="<%= if writable?(function), do: :write, else: :read %>" data-type="<%= @contract_type %>" data-url="<%= smart_contract_path(@conn, :show, @address.hash) %>" data-contract-address="<%= @address.hash %>" data-contract-abi="<%= @contract_abi %>" data-implementation-abi="<%= @implementation_abi  %>" data-chain-id="<%= Explorer.Chain.Cache.NetVersion.get_version() %>">
+        <form class="form-inline" data-function-form data-action="<%= if writable?(function), do: :write, else: :read %>" data-type="<%= @contract_type %>" data-url="<%= smart_contract_path(@conn, :show, Address.checksum(@address.hash)) %>" data-contract-address="<%= @address.hash %>" data-contract-abi="<%= @contract_abi %>" data-implementation-abi="<%= @implementation_abi  %>" data-chain-id="<%= Explorer.Chain.Cache.NetVersion.get_version() %>">
           <input type="hidden" name="function_name" value='<%= function["name"] %>' />
           <input type="hidden" name="method_id" value='<%= function["method_id"] %>' />
 
@@ -80,7 +68,7 @@ to: address_contract_path(@conn, :index, metadata_for_verification.address_hash)
             <% end %>
           <% end %>
 
-          <%= if payable?(function) do %>
+          <%= if Helper.payable?(function) do %>
             <div class="form-group pr-3">
               <input type="text" name="function_input" tx-value class="form-control form-control-sm address-input-sm mt-2" placeholder='value(<%= gettext("ETH")%>)'  />
             </div>
