+++ apps/block_scout_web/lib/block_scout_web/views/tokens/overview_view.ex
@@ -3,6 +3,7 @@ defmodule BlockScoutWeb.Tokens.OverviewView do
 
   alias Explorer.{Chain, CustomContractsHelpers}
   alias Explorer.Chain.{Address, SmartContract, Token}
+  alias Explorer.SmartContract.Helper
 
   alias BlockScoutWeb.{AccessHelpers, CurrencyHelpers, LayoutView}
 
@@ -48,7 +49,7 @@ defmodule BlockScoutWeb.Tokens.OverviewView do
   def smart_contract_with_read_only_functions?(
         %Token{contract_address: %Address{smart_contract: %SmartContract{}}} = token
       ) do
-    Enum.any?(token.contract_address.smart_contract.abi, &(&1["constant"] || &1["stateMutability"] == "view"))
+    Enum.any?(token.contract_address.smart_contract.abi, &Helper.queriable_method?(&1))
   end
 
   def smart_contract_with_read_only_functions?(%Token{contract_address: %Address{smart_contract: nil}}), do: false
