+++ apps/block_scout_web/test/block_scout_web/views/tokens/smart_contract_view_test.exs
@@ -15,6 +15,11 @@ defmodule BlockScoutWeb.SmartContractViewTest do
 
       refute SmartContractView.queryable?(inputs)
     end
+
+    test "returns true if list of inputs is not empty" do
+      assert SmartContractView.queryable?([%{"name" => "argument_name", "type" => "uint256"}]) == true
+      assert SmartContractView.queryable?([]) == false
+    end
   end
 
   describe "writable?" do
@@ -73,128 +78,12 @@ defmodule BlockScoutWeb.SmartContractViewTest do
     end
   end
 
-  describe "payable?" do
-    test "returns true when there is payable function" do
-      function = %{
-        "type" => "function",
-        "stateMutability" => "payable",
-        "payable" => true,
-        "outputs" => [],
-        "name" => "upgradeToAndCall",
-        "inputs" => [
-          %{"type" => "uint256", "name" => "version"},
-          %{"type" => "address", "name" => "implementation"},
-          %{"type" => "bytes", "name" => "data"}
-        ],
-        "constant" => false
-      }
-
-      assert SmartContractView.payable?(function)
-    end
-
-    test "returns true when there is old-style payable function" do
-      function = %{
-        "type" => "function",
-        "payable" => true,
-        "outputs" => [],
-        "name" => "upgradeToAndCall",
-        "inputs" => [
-          %{"type" => "uint256", "name" => "version"},
-          %{"type" => "address", "name" => "implementation"},
-          %{"type" => "bytes", "name" => "data"}
-        ],
-        "constant" => false
-      }
-
-      assert SmartContractView.payable?(function)
-    end
-
-    test "returns false when it is nonpayable function" do
-      function = %{
-        "type" => "function",
-        "stateMutability" => "nonpayable",
-        "payable" => false,
-        "outputs" => [],
-        "name" => "transferProxyOwnership",
-        "inputs" => [%{"type" => "address", "name" => "newOwner"}],
-        "constant" => false
-      }
-
-      refute SmartContractView.payable?(function)
-    end
-
-    test "returns false when there is no function" do
-      function = %{}
-
-      refute SmartContractView.payable?(function)
-    end
-
-    test "returns false when function is nil" do
-      function = nil
-
-      refute SmartContractView.payable?(function)
-    end
-  end
-
-  describe "nonpayable?" do
-    test "returns true when there is nonpayable function" do
-      function = %{
-        "type" => "function",
-        "stateMutability" => "nonpayable",
-        "payable" => false,
-        "outputs" => [],
-        "name" => "transferProxyOwnership",
-        "inputs" => [%{"type" => "address", "name" => "newOwner"}],
-        "constant" => false
-      }
-
-      assert SmartContractView.nonpayable?(function)
-    end
-
-    test "returns true when there is old-style nonpayable function" do
-      function = %{
-        "type" => "function",
-        "outputs" => [],
-        "name" => "test",
-        "inputs" => [%{"type" => "address", "name" => "newOwner"}],
-        "constant" => false
-      }
-
-      assert SmartContractView.nonpayable?(function)
-    end
-
-    test "returns false when it is payable function" do
-      function = %{
-        "type" => "function",
-        "stateMutability" => "payable",
-        "payable" => true,
-        "outputs" => [],
-        "name" => "upgradeToAndCall",
-        "inputs" => [
-          %{"type" => "uint256", "name" => "version"},
-          %{"type" => "address", "name" => "implementation"},
-          %{"type" => "bytes", "name" => "data"}
-        ],
-        "constant" => false
-      }
-
-      refute SmartContractView.nonpayable?(function)
-    end
-
-    test "returns true when there is no function" do
-      function = %{}
-
-      refute SmartContractView.nonpayable?(function)
-    end
-
-    test "returns false when function is nil" do
-      function = nil
-
-      refute SmartContractView.nonpayable?(function)
-    end
-  end
-
   describe "address?" do
+    test "returns true if type equals `address`" do
+      assert SmartContractView.address?("address") == true
+      assert SmartContractView.address?("uint256") == false
+    end
+
     test "returns true when the type is equal to the string 'address'" do
       type = "address"
