+++ apps/explorer/lib/explorer/smart_contract/helper.ex
@@ -1,0 +1,24 @@
+defmodule Explorer.SmartContract.Helper do
+  @moduledoc """
+  SmartContract helper functions
+  """
+
+  def queriable_method?(method) do
+    method["constant"] || method["stateMutability"] == "view"
+  end
+
+  def constructor?(function), do: function["type"] == "constructor"
+
+  def event?(function), do: function["type"] == "event"
+
+  def payable?(function), do: function["stateMutability"] == "payable" || function["payable"]
+
+  def nonpayable?(function) do
+    if function["type"] do
+      function["stateMutability"] == "nonpayable" ||
+        (!function["payable"] && !function["constant"] && !function["stateMutability"])
+    else
+      false
+    end
+  end
+end
